requirements.txt
beautifulsoup4==4.9.3
requests==2.31.0
lxml==4.9.2
findspark==2.0.1
googlemaps==4.10.0
pandas==2.0.2
python-dotenv==1.0.0